requirements.txt
streamlit
kubernetes
openai
python-dotenv
pyyaml
